refactor(platform): move releasePreviousBlob into a shared blob-rotation module
The helper rotates the blob a rewrite replaces (unbind + trash the file
row, enqueue knowledge.release_refs in-tx) for both the agent refresh and
the project-text rewrite; project-text.ts importing it from agent-write.ts
read as one lane reaching into another's internals (PR #3227 review). Same
code, neutral home; both writers import it from there.
